pyproject.toml
[tool.poetry]
name = "python-line-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "python_line_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
twder = "^0.2.2"
apscheduler = "^3.9.1.post1"
flask = "^2.2.2"
gunicorn = "^20.1.0"
line-bot-sdk = "^2.3.0"
lxml = "^4.9.2"
newrelic = "^8.5.0"
pymongo = "^4.3.3"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
redis = "^4.4.0"
beautifulsoup4 = "^4.11.1"
mongoengine = "^0.24.2"
python-dateutil = "^2.8.2"
cachetools = "^5.2.0"
fake-useragent = "^1.1.1"

[tool.poetry.group.test]  # This part can be left out

[tool.poetry.group.test.dependencies]
mongomock = "^4.1.2"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
